Remove stale openhands/README.md and move version.py to app_server
- Delete openhands/README.md (stale documentation)
- Move openhands/version.py to openhands/app_server/version.py
- Create deprecation stub at openhands/version.py that re-exports from app_server
- Update openhands/__init__.py to import from new location
- Update openhands/app_server/app.py to use new location
