+++ openhands/__init__.py
@@ -4,6 +4,6 @@
 __path__ = __import__('pkgutil').extend_path(__path__, __name__)
 
 # Import version information for backward compatibility
-from openhands.version import __version__, get_version
+from openhands.app_server.version import __version__, get_version
 
 __all__ = ['__version__', 'get_version']
